Preserve sub-second precision through the streaming-loop time path
provisional_scan_start_secs() returned i64 via .round() — every volume's
provisional start was truncated to whole seconds before becoming the IDB
scan key, while the worker's later volume_header_time_secs (parsed from
radial headers) was f64 with full precision. Two volumes whose true
starts fell in the same wall-clock second would round to the same i64
and risk colliding under ScanKey::from_secs.

Promote the chain to f64 end-to-end:
  - UnixMillis::from_secs_f64 / as_secs_f64 + ScanKey::from_secs_f64
  - provisional_scan_start_secs() returns f64
  - current_scan_start_secs: f64 in streaming_loop
  - RealtimeResult::ChunkData::timestamp: f64
  - ChunkIngestContext / IngestContext / WorkerOutcome::WorkerError
    .failed_scan_timestamp_secs all carry f64
  - WorkerPool::ingest / ingest_chunk and the underlying send helpers
    take f64 (the JS message format already used f64; the Rust-side cast
    was the only place precision was lost)
  - viz_state.{displayed,prev_sweep_scan}_timestamp: Option<f64>
  - PlaybackManager::find_prev_sweep returns f64 instead of casting via
    'as i64' on the way out

Two boundaries explicitly truncate where sub-second isn't meaningful:
NetworkGroupKey::RealtimeScan (HashMap key, can't be f64; volumes never
share a wall-clock second) and DownloadProgress.in_flight_scans (archive-
derived i64-second tuples). Both sites are commented.

Two new pure-Rust tests pin the precision behaviour.

+++ src/nexrad/decode_worker/mod.rs
@@ -47,12 +47,12 @@ pub struct DecodeWorker {
 
 /// A request queued before the worker was ready.
 enum QueuedRequest {
-    Ingest(RequestId, Vec<u8>, String, i64, String),
+    Ingest(RequestId, Vec<u8>, String, f64, String),
     IngestChunk(
         RequestId,
         Vec<u8>,
         String,
-        i64,
+        f64,
         u32,
         bool,
         bool,
